+++ fatiando/gridder.py
@@ -17,6 +17,12 @@ Create and operate on grids and profiles.
 * :func:`~fatiando.gridder.interp_at`
 * :func:`~fatiando.gridder.extrapolate_nans`
 
+**Padding**
+
+* :func:`~fatiando.gridder.pad_array`
+* :func:`~fatiando.gridder.unpad_array`
+* :func:`~fatiando.gridder.pad_coords`
+
 **Input/Output**
 
 * :func:`~fatiando.gridder.load_surfer`: Read a Surfer grid file and return
@@ -80,7 +86,7 @@ def load_surfer(fname, fmt='ascii'):
             # Read the number of columns (ny) and rows (nx)
             ny, nx = [int(s) for s in ftext.readline().split()]
             shape = (nx, ny)
-            # Read the min/max value of columns/longitue (y direction)
+            # Read the min/max value of columns/longitude (y direction)
             ymin, ymax = [float(s) for s in ftext.readline().split()]
             # Read the min/max value of rows/latitude (x direction)
             xmin, xmax = [float(s) for s in ftext.readline().split()]
@@ -439,3 +445,334 @@ def cut(x, y, scalars, area):
               if x[i] >= xmin and x[i] <= xmax
               and y[i] >= ymin and y[i] <= ymax]
     return [x[inside], y[inside], [s[inside] for s in scalars]]
+
+
+def pad_array(a, npd=None, padtype='OddReflectionTaper'):
+    """
+    Return a padded array of arbitrary dimension.
+
+    The function takes an array of arbitrary dimension and pads it either to
+    the dimensions given by the tuple npd, or to the next power of 2 if npd is
+    not given.
+
+    An odd reflection with a cosine taper is the author's preferred method of
+    padding for Fourier operations.  The odd reflection optimally preserves
+    the frequency content while adding minimal sharp inflections.  The cosine
+    taper is a smooth function which also adds few sharp inflection points.
+
+    .. note:: Requires gridded data of the same dimension as desired
+        (i.e. no flattened arrays; use reshape).
+
+    .. note:: This function returns a deep copy of the original array.
+
+    .. warning:: This function returns the coordinate vectors as a list of
+        arrays.  This means that even for a 1D input array, the returned
+        vector will be in a list of one element.
+
+    Parameters:
+
+    * a : array
+        Array (N-D) to be padded
+    * npd : tuple (optional)
+        Desired shape of new padded array.  If not provided, the nearest
+        power of 2 will be used.
+    * padtype : string (optional)
+        String describing with what to pad the new values. Options:
+
+        [ oddreflectiontaper | oddreflection | reflection | value | lintaper
+        | edge | mean ]
+
+            oddreflectiontaper - Generates odd reflection then tapers to the
+            mean using a cosine function (Default)
+
+            oddreflection - Pads with the odd reflection, with no taper
+
+            reflection - Pads with simple reflection
+
+            lintaper - Linearly tapers to the mean
+
+            value - Numeric value. Input a float or integer directly.
+
+            edge - Uses the edge value as a constant pad
+
+            mean - Uses the mean of the vector along each axis
+
+    Returns:
+
+    * ap : numpy array
+        Padded array. The array core is a deep copy of the original array
+    * nps : list
+        List of tuples containing the number of elements padded onto each
+        dimension.
+
+    Examples:
+
+        >>> z = numpy.array([3, 4, 4, 5, 6])
+        >>> zpad, nps = pad_array(z)
+        >>> zpad
+        array([ 4.4,  3.2,  3. ,  4. ,  4. ,  5. ,  6. ,  4.4])
+        >>> nps
+        [(2, 1)]
+
+        >>> shape = (5, 6)
+        >>> z = numpy.ones(shape)
+        >>> zpad, nps = pad_array(z, padtype='5')
+        >>> zpad.shape
+        (8, 8)
+        >>> nps
+        [(2, 1), (1, 1)]
+
+    """
+
+    # Test to make sure padtype is valid
+    padopts = ['oddreflectiontaper', 'oddreflection', 'reflection',
+               'lintaper', 'edge', 'value', 'mean']
+    # API Note:
+    # If one wishes to add more options to the padding, use the following
+    # checklist to make sure all the sections are consistent.
+    #  [ ] Docstring (2 spots: the options, and the option descriptions
+    #  [ ] padopts - Make sure to add to avoid raising a ValueError
+    #  [ ] if statements below - Finally add an elif to the if statements
+    #                            below. Add a descriptive comment.
+    if str(padtype).lower() not in padopts and not _is_number(padtype):
+        raise ValueError('Pad type not understood')
+    # If npd is not provided, populate with next power of 2
+    npt = []
+    nd = a.ndim
+    if npd is None:
+        for ii in range(0, nd):
+            if nd == 1:
+                npt.append(_nextpow2(len(a)))
+            else:
+                npt.append(_nextpow2(a.shape[ii]))
+    else:
+        et = 'Pad dimensions do not match array dims'
+        if nd == 1:
+            if _is_integer(npd):
+                npt.append(npd)
+            else:
+                if len(npd) != 1:
+                    raise ValueError(et)
+                else:
+                    npt.append(npd)
+        else:
+            if _is_integer(npd):
+                raise ValueError(et)
+            elif len(npd) != a.ndim:
+                raise ValueError(et)
+            else:
+                npt = npd
+        for ii in range(0, len(npt)):
+            if npt[ii] <= a.shape[ii]:
+                raise ValueError(
+                    'Desired padding is less than array ' +
+                    'length along dimension' + str(ii) + '.')
+    # Compute numbers to pad on the left and right side of the array along
+    # each dimension
+    nps = []
+    for ii in range(0, nd):
+        nps.append((int(numpy.ceil((npt[ii] - a.shape[ii])/2.)),
+                    int(numpy.floor((npt[ii] - a.shape[ii])/2.))))
+
+    # If it will be needed, compute the mean
+    meanneeded = ['lintaper', 'oddreflectiontaper']
+    if padtype.lower() in meanneeded:
+        m = numpy.mean(a)
+    # Use numpy's padding routines where possible
+    if _is_number(padtype):
+        # Pad with value
+        ap = numpy.pad(a, nps, mode='constant',
+                       constant_values=(float(padtype), float(padtype)))
+    elif padtype.lower() == 'mean':
+        # Pad with the mean
+        ap = numpy.pad(a, nps, mode='mean')
+    elif padtype.lower() == 'lintaper':
+        # Linearly taper to the mean
+        ap = numpy.pad(a, nps, mode='linear_ramp', end_values=(m, m))
+    elif padtype.lower() == 'edge':
+        # Pad with edge values
+        ap = numpy.pad(a, nps, mode='edge')
+    elif padtype.lower() == 'reflection':
+        # Pad with even reflection
+        ap = numpy.pad(a, nps, mode='reflect',
+                       reflect_type='even')
+    elif padtype.lower() == 'oddreflection':
+        # Pad with odd reflection
+        ap = numpy.pad(a, nps, mode='reflect',
+                       reflect_type='odd')
+    elif padtype.lower() == 'oddreflectiontaper':
+        # Pad with odd reflection and a cosine taper to mean
+        ap = (numpy.pad(a, nps, mode='reflect',
+                        reflect_type='odd') - m)
+        for ii in range(0, nd):
+            ap = numpy.apply_along_axis(_costaper, ii, ap, lp=nps[ii][0],
+                                        rp=nps[ii][1])
+        ap += m
+
+    return ap, nps
+
+
+def unpad_array(a, nps):
+    """
+    Unpads an array using the outputs from pad_array.
+
+    This function takes a padded array and removes the padding from both.
+    Effectively, this is a complement to gridder.cut for when you already
+    know the number of elements to remove.
+
+    .. note:: Unlike :func:`~fatiando.gridder.pad_array`, this returns a slice
+        of the input array.  Therefore, any changes to the padded array will be
+        reflected in the unpadded array.
+
+    Parameters:
+
+    * a : array
+        Array to be un-padded.  Can be of arbitrary dimension.
+    * nps : list
+        List of tuples giving the min and max indices for the cutoff.
+        Identical to nps returned by pad_array
+
+    Returns:
+
+    * b : array
+        Array of same dimension as a, with padding removed
+
+    Examples:
+
+        >>> z = numpy.array([3, 4, 4, 5, 6])
+        >>> zpad, nps = pad_array(z)
+        >>> zpad
+        array([ 4.4,  3.2,  3. ,  4. ,  4. ,  5. ,  6. ,  4.4])
+        >>> zunpad = unpad_array(zpad, nps)
+        >>> zunpad
+        array([ 3.,  4.,  4.,  5.,  6.])
+
+    """
+    o = []
+    for ii in range(0, a.ndim):
+        o.append(slice(nps[ii][0], a.shape[ii] - nps[ii][1]))
+    b = a[o]
+
+    return b
+
+
+def pad_coords(xy, shape, nps):
+    """
+    Pads coordinate vectors.
+
+    Designed to be used in concert with :func:`~fatiando.gridder.pad_array`,
+    this function takes a list of coordinate vectors and pads them using the
+    same discretization.
+
+    .. note:: This function returns a list of arrays in the same format
+        as, for example, :func:`~fatiando.gridder.regular`.  It is a list of
+        flattened meshgrids for each vector in the same order as was input.
+
+    Parameters:
+
+    * xy : list
+        List of arrays of coordinates
+    * shape : tuple
+        Size of original array
+    * nps : list
+        List of tuples containing the number of elements padded onto each
+            dimension (uses output from :func:`~fatiando.gridder.pad_array`).
+
+    Returns:
+
+    * coordspad : list
+        List of padded coordinate arrays
+
+    Examples:
+
+        >>> shape = (5, 6)
+        >>> x, y, z = regular((-10,10,-20,0), shape, z=-25)
+        >>> gz = numpy.zeros(shape)
+        >>> gzpad, nps = pad_array(gz)
+        >>> x.reshape(shape)[:,0]
+        array([-10.,  -5.,   0.,   5.,  10.])
+        >>> y.reshape(shape)[0,:]
+        array([-20., -16., -12.,  -8.,  -4.,   0.])
+        >>> xy = [x, y]
+        >>> N = pad_coords(xy, shape, nps)
+        >>> N[0].reshape(gzpad.shape)[:,0]
+        array([-20., -15., -10.,  -5.,   0.,   5.,  10.,  15.])
+        >>> N[1].reshape(gzpad.shape)[0,:]
+        array([-24., -20., -16., -12.,  -8.,  -4.,   0.,   4.])
+
+    """
+    coords = []
+    d = []
+    coordspad = []
+    for ii in range(0, len(shape)):
+        if type(xy) is not list:
+            coords.append(xy)
+        elif type(xy) is list and len(shape) > 1:
+            coords.append(xy[ii].reshape(shape).transpose().take(0, axis=ii))
+        d.append(coords[ii][1] - coords[ii][0])
+        coordspad.append(_padcvec(coords[ii], nps[ii], d[ii]))
+    M = numpy.meshgrid(*[a for a in tuple(coordspad)[::-1]])
+    N = []
+    for a in M:
+        N.append(a.ravel())
+
+    return N[::-1]
+
+
+def _padcvec(x, n, dx):
+    # Used by padcoords to pad an individual vector based on the number
+    # of points on either side and the point spacing
+    xp = numpy.zeros(len(x) + n[0] + n[1])
+    xp[n[0]:n[0]+len(x)] = x[:]
+    for ii, jj in enumerate(range(0, n[0])[::-1]):
+        xp[ii] = x[0] - ((jj + 1) * dx)
+    for ii, jj in enumerate(range(len(x)+n[0], len(xp))):
+        xp[jj] = x[-1] + (dx * (ii + 1))
+    return xp
+
+
+def _unpadcvec(x, n):
+    # Takes a vector, x, and a tuple, n, and removes n[0] elements from the
+    # left and n[1] to the right
+    return x[n[0]:len(x)-n[1]]
+
+
+def _costaper(a, lp, rp):
+    # This takes an array and applies a cosine taper to each end.
+    # The array has already been deep copied above.  This is by reference only.
+    a[0:lp] = a[0:lp] * _calccostaper(lp)[::-1]
+    a[-rp:] = a[-rp:] * _calccostaper(rp)
+    return a
+
+
+def _calccostaper(ntp):
+    # Used by _costaper to compute a cosine taper from 1 to zero over
+    # ntp points
+    tp = numpy.zeros(ntp)
+    for ii in range(1, ntp + 1):
+        tp[ii-1] = (1.0 + numpy.cos((ii*numpy.pi) / float(ntp))/2.) - 0.5
+    return tp
+
+
+def _nextpow2(ii):
+    # Computes the next power of two
+    buf = numpy.ceil(numpy.log(ii) / numpy.log(2))
+    return int(2**buf)
+
+
+def _is_number(s):
+    # Returns true if s can be cast as a float, false otherwise
+    try:
+        float(s)
+        return True
+    except ValueError:
+        return False
+
+
+def _is_integer(s):
+    # Returns true if s is an integer. Used for testing int/array
+    try:
+        int(s)
+        return True
+    except TypeError:
+        return False